pyproject.toml
[project]
name = "reddit-digest"
version = "0.1.0"
description = "AI-curated Reddit feed that emails you filtered posts"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0",
    "openai>=1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "click>=8.0",
]

[project.scripts]
reddit-digest = "reddit_digest.main:main"
rd = "reddit_digest.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/reddit_digest"]
